feels fast now fix issue with restarting interval
