+++ packages/cli/src/commands/preview/server/__test__/livereload.test.ts
@@ -1,0 +1,13 @@
+import { WATCH_IGNORE } from "../livereload";
+
+jest.useFakeTimers();
+
+describe("livereload", () => {
+  it("ignores the right files when watching", () => {
+    expect(
+      WATCH_IGNORE.test(".mailing/src/emails/node_modules/lodash/fp/rest.js")
+    ).toBe(true);
+    expect(WATCH_IGNORE.test(".next")).toBe(true);
+    expect(WATCH_IGNORE.test("node_modules")).toBe(true);
+  });
+});
