+++ packages/cli/src/commands/preview/server/__test__/start.test.ts
@@ -1,4 +1,4 @@
-import startPreviewServer, { WATCH_IGNORE } from "../start";
+import startPreviewServer from "../start";
 import * as config from "../../../../util/config";
 import http from "http";
 import * as paths from "../../../../util/paths";
@@ -23,13 +23,5 @@ describe("start", () => {
       await startPreviewServer();
     }).rejects.toThrow("previews directory does not exist");
     expect(mockHttpServer).not.toHaveBeenCalled();
-  });
-
-  it("ignores the right files when watching", () => {
-    expect(
-      WATCH_IGNORE.test(".mailing/src/emails/node_modules/lodash/fp/rest.js")
-    ).toBe(true);
-    expect(WATCH_IGNORE.test(".next")).toBe(true);
-    expect(WATCH_IGNORE.test("node_modules")).toBe(true);
   });
 });
