+++ packages/cli/src/commands/util/livereloadUtil.ts
@@ -1,0 +1,4 @@
+// This file is for livereload stuff used on both client and server.
+
+export const LONG_POLLING_INTERVAL = 10000;
+export const SHORT_POLLING_INTERVAL = 10;
